fixed some errors introduced with previous commit
